libcamera: pipeline: vivid: Configure the device
When the configurations have been generated and validated, they can be
applied to a device.

Vivid supports only a single stream, so it directly obtains the first
StreamConfiguration from the CameraConfiguration.

The VIVID catpure device is a V4L2Video device, so we generate a
V4L2DeviceFormat to apply directly to the capture device node.

Note that we convert the libcamera Format stored in
cfg.pixelFormat to a V4L2PixelFormat using V4L2PixelFormat helper. This
currently defaults to the single-planar formats, and should be extended
to support the Multiplanar configuration from the V4L2Device.

[todo Repair the link between the multiplanar configuration of the
      V4L2VideoDevice and the pixel format selection]

Following the call to set the format using the Kernel API, if the format
has been adjusted in any way by the kernel driver, then we have failed
to correctly handle the validation stages, and thus the configure
operation is idendified has having failed.

Finally stream specific data can be directly stored and set as
reflecting the state of the stream.

[NOTE: the cfg.setStream() call here associates the stream to the
StreamConfiguration however that should quite likely be done as part of
the validation process. TBD]

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>