libcamera: pipeline: vivid: Initialise key controls
The VIVID pipeline handler retains state globally of it's controls.
Ensure that when we configure this specific pipeline we set initial
parameters on the device that suit our (specific) needs.

This introduces how controls can be set directly on a device, however
under normal circumstances controls should usually be set from libcamera
controls as part of a request. These are VIVID specific only.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>