py: New event handling
At the moment the Python bindings only handle the requestCompleted
events. But we have others, bufferCompleted and disconnect from the
Camera, and cameraAdded and cameraRemoved from the CameraManager.

This makes all those events available to the users.

The get_ready_requests() method is now deprecated, but available to keep
backward compatibility.

The new event handling works as follows:

The user sets callbacks to the CameraManager or Cameras (e.g.
Camera.buffer_completed). When the event_fd informs of an event, the
user must call CameraManager.dispatch_events() which gets the queued
events and calls the relevant callbacks for each queued event.

Additionally there is CameraManager.discard_events() if the user does
not want to process the events but wants to clear the event queue (e.g.
after stopping the cameras or when exiting the app).

I'm not very happy with this patch. It works fine, but there's a lot of
repetition of almost-the-same code. Perhaps some template magics might
reduce the repetition, but I also fear that it can easily make the code
more difficult to read.

TODO: Documentation.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>