py: Create PyCameraManager
Wrap the CameraManager with a PyCameraManager class and move the related
code inside the new class.

This helps understanding the life times of the used-to-be global
variables, gets rid of static handleRequestCompleted function, and
allows us to simplify the binding code as the more complex pieces are
inside the class.

There should be no user visible functional changes.

Note that attaching to requestCompleted signal is a bit funny, as
apparently the only way to use a lambda with a signal is to provide an
object instance pointer as the first parameter, even if there's really
no instance.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>