py: examples: Add simple-capture-tensorflow.py
Add a simple example showing how to use opencv and tensorflow with
libcamera.

This is only a simple example, and it has the same issue as all the
other examples: pixel format conversions. Each pixel format requires
custom conversion code, and I have implemented the conversion only for a
few of them.

Also, this is almost identical to simple-capture.py. Normally it would
make sense to add common helpers, but so far I have stayed away from it
to keep the examples independent and in a single file. Making the
examples "too fancy" will easily make them more difficult to understand
for a newcomer.

Note: I took the datafiles from picamera2 repository. I could not find
the license for them.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>