py: Improve stub type generation for PyCameraEvent
We store the camera, request and fb fields in PyCameraEvent as
py::objects, so that we keep the refs and keep-alives. When we
return py::objects to the Python side, they, obviously, show up
as "objects" when the stub generation is looking at the fields.

Fix this by manually casting the py::objects to the correct C++
counterparts, which causes pybind11 to assign the correct types for the
properties, helpind stubgen to generate correct types.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>